utils/audio_tools.py
import logging
import os
import time
import requests
from typing import Optional, Tuple
from utils.error_handling import retry

logger = logging.getLogger(__name__)

@retry(max_attempts=3, initial_delay=2.0)
def download_audio(audio_url: str, output_path: Optional[str] = None) -> str:
    """
    Download audio from a URL.
    
    Args:
        audio_url: URL to the audio file
        output_path: Optional path to save the audio file
        
    Returns:
        Path to the downloaded audio file
    """
    logger.info(f"Downloading audio from: {audio_url}")
    
    if not output_path:
        # Create a unique filename
        timestamp = int(time.time())
        filename = f"audio_{timestamp}.mp3"
        output_path = os.path.join("data", filename)
    
    # Create directory if it doesn't exist
    os.makedirs(os.path.dirname(output_path), exist_ok=True)
    
    # In a real implementation, you would download the file
    # For now, this is a placeholder
    """
    response = requests.get(audio_url, stream=True)
    response.raise_for_status()
    
    with open(output_path, 'wb') as f:
        for chunk in response.iter_content(chunk_size=8192):
            f.write(chunk)
    """
    
    # Create a placeholder file
    with open(output_path, 'w') as f:
        f.write(f"Placeholder for audio downloaded from {audio_url}")
    
    logger.info(f"Audio downloaded to: {output_path}")
    
    return output_path

def get_audio_duration(audio_path: str) -> float:
    """
    Get the duration of an audio file in seconds.
    
    Args:
        audio_path: Path to the audio file
        
    Returns:
        Duration in seconds
    """
    # In a real implementation, you would use a library like librosa or pydub
    # For now, this is a placeholder
    """
    from pydub import AudioSegment
    audio = AudioSegment.from_file(audio_path)
    return len(audio) / 1000.0  # Convert milliseconds to seconds
    """
    
    # Return a placeholder duration
    logger.info(f"Getting duration for audio: {audio_path}")
    return 180.0  # 3 minutes

def analyze_audio(audio_path: str) -> dict:
    """
    Analyze audio to extract features like BPM, key, etc.
    
    Args:
        audio_path: Path to the audio file
        
    Returns:
        Dictionary with audio features
    """
    # In a real implementation, you would use a library like librosa
    # For now, this is a placeholder
    """
    import librosa
    
    y, sr = librosa.load(audio_path)
    
    # Extract tempo (BPM)
    tempo, _ = librosa.beat.beat_track(y=y, sr=sr)
    
    # Extract key
    chroma = librosa.feature.chroma_cqt(y=y, sr=sr)
    key = librosa.key.estimate_key(chroma)
    
    return {
        "bpm": tempo,
        "key": key,
        "duration": len(y) / sr
    }
    """
    
    # Return placeholder features
    logger.info(f"Analyzing audio: {audio_path}")
    return {
        "bpm": 120,
        "key": "C major",
        "duration": 180.0,
        "energy": 0.8,
        "danceability": 0.7
    }

def split_audio(audio_path: str, segments: list) -> list:
    """
    Split audio into segments.
    
    Args:
        audio_path: Path to the audio file
        segments: List of (start_time, end_time) tuples in seconds
        
    Returns:
        List of paths to the split audio files
    """
    # In a real implementation, you would use a library like pydub
    # For now, this is a placeholder
    """
    from pydub import AudioSegment
    
    audio = AudioSegment.from_file(audio_path)
    output_paths = []
    
    for i, (start, end) in enumerate(segments):
        # Convert to milliseconds
        start_ms = start * 1000
        end_ms = end * 1000
        
        # Extract segment
        segment = audio[start_ms:end_ms]
        
        # Save segment
        output_path = f"{os.path.splitext(audio_path)[0]}_segment_{i}.mp3"
        segment.export(output_path, format="mp3")
        
        output_paths.append(output_path)
    
    return output_paths
    """
    
    # Return placeholder paths
    logger.info(f"Splitting audio: {audio_path} into {len(segments)} segments")
    
    output_paths = []
    for i, (start, end) in enumerate(segments):
        output_path = f"{os.path.splitext(audio_path)[0]}_segment_{i}.mp3"
        
        # Create a placeholder file
        with open(output_path, 'w') as f:
            f.write(f"Placeholder for audio segment {i} ({start}s to {end}s)")
        
        output_paths.append(output_path)
    
    return output_paths
