requirements.txt
# Core dependencies
python-dotenv==1.0.0
supabase==1.0.3
openai==1.3.0
requests==2.31.0
runwayml==0.7.0  # RunwayML SDK for video generation

# Video processing (uncomment as needed)
moviepy==1.0.3
# ffmpeg-python==0.2.0

# Audio processing (uncomment as needed)
pydub==0.25.1
# librosa==0.10.0

# Retry and error handling
tenacity==8.2.2

# Testing
pytest==7.3.1
pytest-mock==3.10.0

# Type checking
mypy==1.3.0
types-requests==2.31.0.1
